Probe the kernel through a ringless io_uring_register
RT-005 in progress. `./run.sh` now prints, after RT-004's `1A2B3C`:

    ops 65 reg 38 feat 262143

All three are checkable against the pinned tree. 65 is IORING_OP_LAST, so
valid request opcodes are 0..64. 38 is IORING_REGISTER_LAST, one past
IORING_REGISTER_BPF_FILTER. And 262143 is 0x3FFFF — bits 0 through 17 with
nothing above, where bit 17 is IORING_FEAT_NO_IOWAIT (io_uring.h:647), the
newest feature this kernel defines. So it supports every feature it can name,
and NO_IOWAIT is confirmed present rather than assumed from a version number.

The probe runs before any ring exists. io_uring_register with fd == -1 takes
the blind path (register.c:1031), which is the point: a later setup failure can
be read against what the kernel already said it supports. Four constraints from
query.c, none of which report themselves usefully — nr_args must be 0
(query.c:133), three header fields must be zero on input (query.c:85), the
kernel copies *from* the result buffer before writing it (query.c:88), and a
zero syscall return does not mean success, since per-entry status lands in
hdr.result (query.c:112).

Scaffolding for the rest: ring.h and ring.c with the five ring operations, each
stubbed with __builtin_trap() rather than a plausible return value, so an
unwritten one stops at a known address instead of failing somewhere unrelated.
struct rt_ring holds only what cannot be recomputed — entry counts are mask + 1
and are not stored, while params.features is unrecoverable once setup returns
and so is. Head and tail are _Atomic unsigned *, which on aarch64 costs nothing
(seq_cst and acquire/release are the same ldar/stlr) but makes a plain access
correct instead of silently wrong.

syscall.h gains sys2 and sys4 with typed wrappers for all three io_uring
syscalls, and put_str/put_dec in main.c format into a buffer and emit once,
rather than one syscall per character.

Build: the uapi headers do not compile under -pedantic — <linux/io_uring.h>
trips -Wzero-length-array on the SQE's cmd[0], plus -Wgnu-empty-struct and
-Wflexible-array-extensions on __DECLARE_FLEX_ARRAY. Moving the include path
from -I to -isystem suppresses diagnostics inside them while leaving -pedantic
aimed at our code; relaxing the flag instead would have disabled those checks
everywhere.

size_t audit across the tree: every value that is a count of bytes in an object
now says so — raw_write, sys_mmap, sys_mprotect and sys_io_uring_enter lengths,
rt_task.stack_len, the stack geometry constants. Register widths, printable
numbers and mmap's file offset are deliberately left alone; they are not sizes.
All translation units are clean under -Wconversion.

+++ src/task.c
@@ -21,9 +21,9 @@
 #include "syscall.h"
 
 /* Stack geometry. One guard page below the usable region. */
-constexpr unsigned long RT_PAGE_SIZE = 4096;
-constexpr unsigned long RT_GUARD_SIZE = RT_PAGE_SIZE;
-constexpr unsigned long RT_STACK_SIZE = 64 * 1024;
+constexpr size_t RT_PAGE_SIZE = 4096;
+constexpr size_t RT_GUARD_SIZE = RT_PAGE_SIZE;
+constexpr size_t RT_STACK_SIZE = 64 * 1024;
 
 /* Both properties are accidents of the numbers above, not guarantees, and
  * neither fails visibly. mmap promises page alignment and nothing more. */
